requirements.txt
pyTelegramBotAPI==4.29.1
python-telegram-bot==22.4
python-dotenv==1.1.1